# app/api/sla/route.ts
import { NextResponse } from "next/server";
import { getUser, createAdminClient, getTenant } from "@/lib/supabaseServer";

// GET /api/sla — fetch SLA metrics for current tenant
export async function GET(req: Request) {
  const user = await getUser();
  if (!user) return NextResponse.json({ error: "Unauthorized" }, { status: 401 });

  const tenant = await getTenant();
  if (!tenant) return NextResponse.json({ error: "Tenant not found" }, { status: 404 });

  const { searchParams } = new URL(req.url);
  const limit = Math.min(Number(searchParams.get("limit") ?? "200"), 500);

  const supabase = await createAdminClient();
  const { data: metrics, error } = await supabase
    .from("sla_metrics")
    .select("*")
    .eq("tenant_id", tenant.id)
    .order("incident_start", { ascending: false })
    .limit(limit);

  if (error) return NextResponse.json({ error: error.message }, { status: 500 });
  return NextResponse.json({ metrics });
}